Refactor App.jsx and fix loading bug
